Delete support for the Conkeror web browser
The last release of Conkeror was in 2017.  The project has been actively
recommending people not to use it since 2019 due to security concerns,
and it was removed from Debian in the same year.  Emacs's browse-url
support for Conkeror has been obsolete since in Emacs 28.1.

* lisp/net/browse-url.el (browse-url-conkeror-new-window-is-buffer):
(browse-url-conkeror-program, browse-url-conkeror-arguments): Delete
variables.
(browse-url-conkeror): Delete function.