Re-enable caching for 'M-x man' completion on macOS
On macOS, getting the completion results when you type 'M-x man RET' can
take up to several seconds, even on a reasonably modern MacBook Pro,
probably due to the lack of man-db and the need to process the output.

Caching of "man -k ^" completion results was added in 2009 due to
perceived slowness on some systems, but was unfortunately effectively
disabled again in 2013.  Let's re-enable it on macOS.

* lisp/man.el (Man-cache-completion-results-flag): New variable
set to t by default on macOS, and nil elsewhere.
(man): Cache completion results when above new variable is t.