New test for finding C header files
The assumption here is that if there is a C compiler, there is also a
math.h header somewhere.  The test should fail if we can't find that
file, and hopefully that will provoke users to create bug reports.

Let's see how far we can take this idea; we might have to give up and
disable the test in some configurations.  But doing that now seems
premature, even if we had a list of affected systems (which we don't).

* lisp/man.el (man--find-header-file): Factor out new function...
(Man-view-header-file): ...from here.
* test/lisp/man-tests.el (man-tests-find-header-file): New test.