* lisp/man.el (Man-fit-to-window): Don't update on tty with MANWIDTH.
Check if 'Man-columns' is non-nil before trying to reformat
the manpage buffer.  The value of the varible 'Man-columns'
is nil on a non-window system where "MANWIDTH" or "COLUMNS"
is defined.  And the return value of the function 'Man-columns'
is meaningless in this case anyway.  So no need to react
to changes in window width (bug#74924).