Mark kermit.el as obsolete
While Kermit is still developed, this Emacs support library is
essentially unchanged since it was added to Emacs, which seems to have
been in 1988.  It is also based on the old pre-Emacs 20 shell mode, so
it's not clear if it even works, and much less if it is relevant with
Kermit from this century.

There is also this modern alternative, from 1994, which seems far
more complete: https://www.kermitproject.org/archive.html#emacs

The bug tracker has stayed silent, so if anyone is still using it, they
will have to report a bug and ask us to unobsolete it.

* lisp/obsolete/kermit.el: Add Obsolete-since header.  Add link to
alternative library at kermitproject.org.